pyproject.toml
[tool.poetry]
name = "city-weather-status"
version = "0.1.0"
description = ""
authors = ["Lea <lea.peligry@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.36.0"
folium = "^0.17.0"
streamlit_folium = "^0.21.1"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
